feat: models 패키지 생성 및 토크나이저 작업 로직 파일 분리
